refactor: split off protocol fees function from WeightedMath
