+++ pkg/pool-utils/contracts/test/MockInvariantGrowthProtocolSwapFees.sol
@@ -1,0 +1,34 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+pragma solidity ^0.7.0;
+
+import "../InvariantGrowthProtocolSwapFees.sol";
+
+contract MockInvariantGrowthProtocolSwapFees {
+    function calculateDueProtocolFees(
+        uint256 invariantGrowthRatio,
+        uint256 previousSupply,
+        uint256 currentSupply,
+        uint256 protocolSwapFeePercentage
+    ) external pure returns (uint256) {
+        return
+            InvariantGrowthProtocolSwapFees._calcDueProtocolFees(
+                invariantGrowthRatio,
+                previousSupply,
+                currentSupply,
+                protocolSwapFeePercentage
+            );
+    }
+}
