+++ pkg/pool-weighted/contracts/WeightedMath.sol
@@ -406,77 +406,6 @@ library WeightedMath {
         return amountsOut;
     }
 
-    function _calcDueProtocolSwapFeeBptAmount(
-        uint256 invariantGrowthRatio,
-        uint256 previousSupply,
-        uint256 currentSupply,
-        uint256 protocolSwapFeePercentage
-    ) internal pure returns (uint256 protocolFeeAmount) {
-        // Joins and exits are symmetrical; for simplicity, we consider a join, where the invariant and supply
-        // both increase.
-
-        // |-------------------------|-- original invariant * invariantGrowthRatio
-        // |   increase from fees    |
-        // |-------------------------|-- original invariant * supply growth ratio (fee-less invariant)
-        // |                         |
-        // | increase from balances  |
-        // |-------------------------|-- original invariant
-        // |                         |
-        // |                         |  |------------------|-- currentSupply
-        // |                         |  |    BPT minted    |
-        // |                         |  |------------------|-- previousSupply
-        // |   original invariant    |  |  original supply |
-        // |_________________________|  |__________________|
-        //
-        // If the join is proportional, the invariant and supply will likewise increase proportionally,
-        // so the growth ratios (invariantGrowthRatio / supplyGrowthRatio) will be equal. In this case, we do not charge
-        // any protocol fees.
-
-        uint256 supplyGrowthRatio = currentSupply.divDown(previousSupply);
-        if ((supplyGrowthRatio == invariantGrowthRatio) || (protocolSwapFeePercentage == 0)) return 0;
-
-        // If the join is non-proportional, the supply increase will be proportionally less than the invariant increase,
-        // since the BPT minted will be based on fewer tokens (because swap fees are not included). So the supply growth
-        // is due entirely to the balance changes, while the invariant growth also includes swap fees.
-        //
-        // To isolate the amount of increase by fees then, we multiply the original invariant by the supply growth
-        // ratio to get the "feeless invariant". The difference between the final invariant and this value is then
-        // the amount of the invariant due to fees, which we convert to a percentage by normalizing against the
-        // final invariant. This is expressed as the expression below:
-        //
-        // invariantGrowthFromFees = currentInvariant - supplyGrowthRatio * previousInvariant
-        //
-        // We then divide through by current invariant so the LHS can be identified as the fraction of the pool which
-        // is made up of accumulated swap fees.
-        //
-        // swapFeesPercentage = 1 - supplyGrowthRatio * previousInvariant / currentInvariant
-        //
-        // We then define `invariantGrowthRatio` in a similar fashion to `supplyGrowthRatio` to give the result:
-        //
-        // swapFeesPercentage = 1 - supplyGrowthRatio / invariantGrowthRatio
-        //
-        // Using this form allows us only consider the ratios of the two invariants rather than absolute values,
-        // a useful property as this is sometimes easier than calculating the full invariant twice.
-
-        uint256 nonSwapFeesPercentage = supplyGrowthRatio.divDown(invariantGrowthRatio);
-
-        // This shouldn't occur, but if the BPT supply increases faster / decreases slower than the invariant then
-        // pay no protocol fees.
-        if (nonSwapFeesPercentage >= FixedPoint.ONE) return 0;
-
-        uint256 swapFeesPercentage = FixedPoint.ONE - nonSwapFeesPercentage;
-
-        // We then multiply by the protocol swap fee percentage to get the fraction of the pool which the protocol
-        // should own once fees have been collected.
-        uint256 protocolOwnershipPercentage = swapFeesPercentage.mulDown(protocolSwapFeePercentage);
-
-        // The percentage of the Pool the protocol will own after minting is given by:
-        // `protocol percentage = to mint / (current supply + to mint)`.
-        // Solving for `to mint`, we arrive at:
-        // `to mint = current supply * protocol percentage / (1 - protocol percentage)`.
-        return currentSupply.mulDown(protocolOwnershipPercentage).divDown(protocolOwnershipPercentage.complement());
-    }
-
     /**
      * @dev Calculate the amount of BPT which should be minted when adding a new token to the Pool.
      *
